Script to create an SCCS managed tzfiles dir
This script will copy ../elsie/tzdata2011l.tar.gz and
../elsie/tzcode2011i.tar.gz into sccs.tz/ and create SCCS version info
using the current SID as determined by the @(#) string in the file.

Not that it's a little kludgy since some older @(#) strings remain - the
released files obviously do not have the %W% markers that SCCS uses and
these have to be put back in.